pyproject.toml
[project]
name = "news-chatbot"
version = "0.1.0"
description = ""
authors = [
    { name = "abhiudaytomar", email = "abhiuday.prospr@gmail.com" },
]
readme = "README.md"
requires-python = "^3.13"
dependencies = ["weaviate-client (>=4.16.10,<5.0.0)", "requests (>=2.32.5,<3.0.0)", "python-dotenv (>=1.1.1,<2.0.0)", "pandas (>=2.3.2,<3.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "news_chatbot", from = "src" }]
